requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
plotly>=5.20.0
tushare>=1.2.89
